7090158: Networking Libraries don't build with javac -Werror 7125055: ContentHandler.getContent API changed in error
Minor changes to networking java files to remove warnings

+++ jdk/src/share/classes/sun/net/www/http/KeepAliveStream.java
@@ -85,7 +85,7 @@ class KeepAliveStream extends MeteredStream implements Hurryable {
         // NOTE: Don't close super class
         try {
             if (expected > count) {
-                long nskip = (long) (expected - count);
+                long nskip = expected - count;
                 if (nskip <= available()) {
                     long n = 0;
                     while (n < nskip) {
